Make operations by interacting with the calculator
The display is updated whenever there's an interaction with the buttons. All was done with click eventListeners and by storing values in variables for the first and second numbers. The separation is made by the operatorValue variable, which is called when there's a click in any operator. Ifs do the work.