Make possible to string several operations
Another if was added below the ones that verify single operations. The new one verifies if all the parameters are filled, and if so, it calls the operate() function, and the result of the call is assigned to the firstNumber variable. Every pair of numbers is assigned to it. By making this, the operation can go on as the user wants.