Delete digits div
For the sake of css manipulation and concision, the for loop which creates the digit buttons was made directly inside the options-and-digits div, so there is no conflict between the options and the digits while styling. They are brothers, but don't share the same class, by which the digits are manipulated in javascript.